requirements.txt
httpx==0.22.0
loguru==0.6.0
Pillow==9.5.0
playwright==1.33.0
pydantic==1.9.0
PyYAML==6.0
